refactor(cli): extract run + interactive drivers into commands/run.rs
Move run_firmware (+ arm/riscv/esp32 fast-boot paths) and run_interactive
(+ arm/riscv/xtensa gdb/dap drivers) into commands/run.rs. They only share
emit_error (now pub(crate)); build_system_bus/run_start/resolve_chip_descriptor_path
were already pub(crate).

Pure move — build + clippy clean.
