+++ crates/cli/src/commands/mod.rs
@@ -1,5 +1,6 @@
 //! CLI subcommand handlers, split out of `main.rs`.
 
 pub mod asset;
+pub mod run;
 pub mod snapshot;
 pub mod test;
